Refactor fio and uperf to create a base image and use it
We create a `scale-ci-base` image which contains all the common parts between
`scale-ci-fio` and `scale-ci-uperf` previously, and then make simple docker
files for those two images adding only the `fio` and `uperf` RPMs respectively.

The `scale-ci-base` image leverages the new `pbench-agent-tools-centos-7` image
which properly installs the pbench-agent RPM on a CentOS 7 base image, along
with the RPMs for all the tools.